refactor(YALB-639): namespaced page-layouts and page-examples
