+++ .storybook/preview.js
@@ -9,6 +9,7 @@ import '../components/00-tokens/colors/cl-colors.scss';
 import '../components/00-tokens/cl-base.scss';
 import '../components/00-tokens/typography/cl-typography.scss';
 import '../components/00-tokens/effects/cl-effects.scss';
+import '../components/04-page-layouts/placeholder/cl-placeholder.scss';
 
 // If in a Drupal project, it's recommended to import a symlinked version of drupal.js.
 import './_drupal.js';
