+++ components/04-page-layouts/_yds-base.twig
@@ -11,7 +11,7 @@
     <main {{ bem('main-content') }}>
       {# Main Content #}
       {% block page__content %}
-        {% include "@page-layouts/placeholder/placeholder.twig" with {
+        {% include "@page-layouts/placeholder/yds-placeholder.twig" with {
           placeholder: 'Main Content',
         } %}
       {% endblock %}
