--- components/04-page-layouts/full-width.twig
@@ -1,1 +1,0 @@
-{% extends "@page-layouts/_base.twig" %}
