+++ components/04-page-layouts/yds-full-width.twig
@@ -1,0 +1,1 @@
+{% extends "@page-layouts/_yds-base.twig" %}
