+++ components/05-page-examples/events/event-list.twig
@@ -1,4 +1,4 @@
-{% extends "@page-layouts/full-width.twig" %}
+{% extends "@page-layouts/yds-full-width.twig" %}
   {% block page__header %}
     {% include "@organisms/site-header/yds-site-header.twig" %}
     {% include "@organisms/menu/breadcrumbs/yds-breadcrumbs.twig" with {
