+++ components/05-page-examples/events/events.stories.js
@@ -1,5 +1,5 @@
 // Shared Storybook args.
-import argTypes, { eventArgTypes } from '../../04-page-layouts/page-args';
+import argTypes, { eventArgTypes } from '../../04-page-layouts/cl-page-args';
 
 // Twig files.
 import eventPageTwig from './event-page.twig';
