+++ components/05-page-examples/miscellaneous/accordion-page.twig
@@ -1,4 +1,4 @@
-{% extends "@page-layouts/full-width.twig" %}
+{% extends "@page-layouts/yds-full-width.twig" %}
   {% block page__header %}
     {% include "@molecules/alert/yds-alert.twig" %}
     {% include "@organisms/site-header/yds-site-header.twig" %}
