+++ components/05-page-examples/miscellaneous/miscellaneous-page.stories.js
@@ -1,5 +1,5 @@
 // Shared Storybook args.
-import argTypes from '../../04-page-layouts/page-args';
+import argTypes from '../../04-page-layouts/cl-page-args';
 
 // Twig files.
 import accordionPageTwig from './accordion-page.twig';
