+++ components/05-page-examples/standard-pages/standard-page-with-banner.twig
@@ -1,4 +1,4 @@
-{% extends "@page-layouts/full-width.twig" %}
+{% extends "@page-layouts/yds-full-width.twig" %}
   {% block page__content %}
     {% include "@molecules/banner/cta/yds-cta-banner.twig" %}
     {% include "@page-examples/_intro-content-examples.twig" %}
