+++ components/05-page-examples/standard-pages/standard-page.twig
@@ -1,4 +1,4 @@
-{% extends "@page-layouts/full-width.twig" %}
+{% extends "@page-layouts/yds-full-width.twig" %}
   {% block page__content %}
     {% include "@molecules/page-title//yds-page-title.twig" %}
     {% include "@page-examples/_intro-content-examples.twig" %}
